Refactor packages & comment out skills
