+++ src/main/java/me/florial/utils/Message.java
@@ -3,7 +3,6 @@ package me.florial.utils;
 
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.event.HoverEvent;
-import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.entity.Player;
